+++ package/Runtime/ResourceProviders/ResourceProviderBase.cs
@@ -8,12 +8,26 @@ namespace UnityEngine.ResourceManagement
     /// </summary>
     public abstract class ResourceProviderBase : IResourceProvider
     {
-        protected ResourceProviderBase() { }
+        protected string m_providerId;
 
+        protected ResourceProviderBase() { }
         /// <inheritdoc/>
         public virtual string ProviderId
         {
-            get { return GetType().FullName; }
+            get
+            {
+                if(string.IsNullOrEmpty(m_providerId))
+                    m_providerId = GetType().FullName;
+
+                return m_providerId;
+            }
+        }
+
+        /// <inheritdoc/>
+        public virtual bool Initialize(string id, string data)
+        {
+            m_providerId = id;
+            return !string.IsNullOrEmpty(m_providerId);
         }
 
         /// <inheritdoc/>
@@ -35,7 +49,6 @@ namespace UnityEngine.ResourceManagement
         public abstract IAsyncOperation<TObject> Provide<TObject>(IResourceLocation location, IAsyncOperation<IList<object>> loadDependencyOperation)
         where TObject : class;
 
-        /// <inheritdoc/>
         public virtual bool Release(IResourceLocation location, object asset)
         {
             if (location == null)