+++ package/Runtime/ResourceProviders/Simulation/VirtualAssetBundleProvider.cs
@@ -3,31 +3,45 @@ using System.Collections.Generic;
 
 namespace UnityEngine.ResourceManagement
 {
-    internal class VirtualAssetBundleProvider : ResourceProviderBase
+    /// <summary>
+    /// Simulates the loading behavior of an asset bundle.  Internally it uses the AssetDatabase API.  This provider will only work in the editor.
+    /// </summary>
+    public class VirtualAssetBundleProvider : ResourceProviderBase
     {
-        VirtualAssetBundleManager m_assetBundleManager;
-        string m_providerId;
-        public VirtualAssetBundleProvider(VirtualAssetBundleManager abm, string provId)
+        VirtualAssetBundleManager m_manager;
+        /// <summary>
+        /// Construct a new VirtualAssetBundleProvider object;
+        /// </summary>
+        public VirtualAssetBundleProvider()
         {
-            m_assetBundleManager = abm;
-            m_providerId = provId;
+            m_providerId = typeof(AssetBundleProvider).FullName;
         }
 
-        public override string ProviderId { get { return m_providerId; } }
+        /// <summary>
+        /// Construct a new VirtualAssetBundleProvider object with a specific manager and id
+        /// </summary>
+        /// <param name="mgr">The VirtualAssetBundleManager to use.</param>
+        /// <param name="providerId">The provider id.</param>
+        public VirtualAssetBundleProvider(VirtualAssetBundleManager mgr, string providerId)
+        {
+            m_providerId = providerId;
+            m_manager = mgr;
+        }
+
         internal class InternalOp<TObject> : InternalProviderOperation<TObject>
             where TObject : class
         {
-            public VirtualAssetBundleManager assetBundleManager;
             System.Action<IAsyncOperation<IList<object>>> action;
             IAsyncOperation m_dependencyOperation;
             IAsyncOperation<VirtualAssetBundle> m_requestOperation;
+            VirtualAssetBundleManager m_manager;
             public InternalOp()
             {
                 action = (op) =>
                 {
                     if (op == null || op.Status == AsyncOperationStatus.Succeeded)
                     {
-                        (m_requestOperation = assetBundleManager.LoadAsync(Context as IResourceLocation)).Completed += (bundleOp) =>
+                        (m_requestOperation = m_manager.LoadAsync(Context as IResourceLocation)).Completed += (bundleOp) =>
                         {
                             SetResult(bundleOp.Result as TObject);
                             OnComplete();
@@ -57,9 +71,10 @@ namespace UnityEngine.ResourceManagement
                 }
             }
 
-            public InternalProviderOperation<TObject> Start(IResourceLocation location, IAsyncOperation<IList<object>> loadDependencyOperation)
+            public InternalProviderOperation<TObject> Start(VirtualAssetBundleManager mgr, IResourceLocation location, IAsyncOperation<IList<object>> loadDependencyOperation)
             {
                 Context = location;
+                m_manager = mgr;
                 m_result = null;
                 m_requestOperation = null;
                 m_dependencyOperation = loadDependencyOperation;
@@ -73,22 +88,30 @@ namespace UnityEngine.ResourceManagement
             internal override TObject ConvertResult(AsyncOperation op) { return null; }
         }
 
+        VirtualAssetBundleManager GetManager()
+        {
+            if (m_manager == null)
+                m_manager = VirtualAssetBundleManager.CreateManager(ResourceManager.OnResolveInternalId);
+            return m_manager;
+        }
+
+        /// <inheritdoc/>
         public override IAsyncOperation<TObject> Provide<TObject>(IResourceLocation location, IAsyncOperation<IList<object>> loadDependencyOperation)
         {
             if (location == null)
                 throw new System.ArgumentNullException("location");
             var operation = AsyncOperationCache.Instance.Acquire<InternalOp<TObject>>();
-            operation.assetBundleManager = m_assetBundleManager;
-            return operation.Start(location, loadDependencyOperation);
+            return operation.Start(GetManager(), location, loadDependencyOperation);
         }
 
+        /// <inheritdoc/>
         public override bool Release(IResourceLocation location, object asset)
         {
             if (location == null)
                 throw new System.ArgumentNullException("location");
             if (asset == null)
                 throw new System.ArgumentNullException("asset");
-            return m_assetBundleManager.Unload(location);
+            return GetManager().Unload(location);
         }
     }
 }