+++ package/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.4.0-preview] - 2018-10-01
+ - Exposed CachedProvider.Settings object that can be used to set up the cached provider during the build.
+ - Fixed bug where IAsyncOperations were being released prematurely when passed to callback parameters
+
 ## [2.3.0-preview] - 2018-08-20
  - Minor code formatting and cleanup
  - Aditonal API documentation